chore(testing): Add tests for bootstrap
